Add context implant function
The alchemy.implant() function allows us to take a class instance
and implant certain properties into its context *without* changing
the original instance's context.

This is really useful for controller & model methods and their
respective components & behaviours.

Though it's a bit harder to maintain,
this is a much faster method compared to creating a new class
instance on every request.