Fix DbQuery code to allow field paths
Now you are able to enter conditions like 'settings.model' in your
find options.

This used to fail, as it would think 'settings' was the model you
wanted to look in.

Now only the entry before a first point will be used as the model
if it is already in the correct modelName form (CamelCased)