✨ add schema and parameter definitions to routes
Adds the ability to get the schema and parameter definitions for routes that have parameters.
The schema property is added to the Route class to return a schema with all the fields for a route, and the param_definitions property returns an object that maps each parameter's name to its definition.
A document can now also be supplied whole as a route parameter